+++ functions/Actions.php
@@ -384,6 +384,16 @@ $RosarioActions = [];
 	 * @since 8.7
 	 */
 	/*'ProgramFunctions/SendEmail.fnc.php|send_error' => array(),
+
+	/**
+	 * Widgets init
+	 * Add your add-on custom widgets to the $myextra['Widgets'] var:
+	 * $myextra['Widgets']['Addon_Name'] = [ 'widget_1', 'widget_2' ];
+	 * And load your custom \Addon_Name\Widget_[widget_name] class(es).
+	 *
+	 * @since 10.4
+	 */
+	/*'functions/Widgets.fnc.php|widgets_init' => array(),
 );*/
 
 /**