+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 10.4
 - JS Only show laoding spinner if file input has selected files in warehouse.js
 - Add File Attached Input for existing Fees/Payments in StudentFees.php, StudentPayments.php & Student_Billing/functions.inc.php
 - Add File Attached Input for existing Salaries/Staff Payments/Incomes/Expenses in Expenses.php, Incomes.php, Salaries, StaffPayments.php & Accounting/functions.inc.php
+- Add-ons can add their custom Widgets in classes/core/Widgets.php
 
 Changes in 10.3.3
 -----------------