+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 10.4
 - Add File Attached Input for existing Salaries/Staff Payments/Incomes/Expenses in Expenses.php, Incomes.php, Salaries, StaffPayments.php & Accounting/functions.inc.php
 - Add-ons can add their custom Widgets in classes/core/Widgets.php
 - Add Widgets init action hook in Actions.php & Widgets.fnc.php
+- Fix SQL check student is actually enrolled in Enrollment.fnc.php
 
 Changes in 10.3.3
 -----------------